+++ x/merkledb/codec_test.go
@@ -16,6 +16,8 @@ import (
 )
 
 func FuzzCodecBool(f *testing.F) {
+	f.Add([]byte{0})
+	f.Add([]byte{1})
 	f.Fuzz(
 		func(
 			t *testing.T,
@@ -28,7 +30,7 @@ func FuzzCodecBool(f *testing.F) {
 			startLen := reader.Len()
 			got, err := codec.decodeBool(reader)
 			if err != nil {
-				return
+				t.SkipNow()
 			}
 			endLen := reader.Len()
 			numRead := startLen - endLen
@@ -56,7 +58,7 @@ func FuzzCodecInt(f *testing.F) {
 			startLen := reader.Len()
 			got, err := codec.decodeInt(reader)
 			if err != nil {
-				return
+				t.SkipNow()
 			}
 			endLen := reader.Len()
 			numRead := startLen - endLen
@@ -84,7 +86,7 @@ func FuzzCodecSerializedPath(f *testing.F) {
 			startLen := reader.Len()
 			got, err := codec.decodeSerializedPath(reader)
 			if err != nil {
-				return
+				t.SkipNow()
 			}
 			endLen := reader.Len()
 			numRead := startLen - endLen
@@ -97,7 +99,8 @@ func FuzzCodecSerializedPath(f *testing.F) {
 			require.Equal(b[:numRead], bufBytes)
 
 			clonedGot := got.deserialize().Serialize()
-			require.Equal(got, clonedGot)
+			require.Equal(got.Value, clonedGot.Value)
+			require.Equal(got.NibbleLength, clonedGot.NibbleLength)
 		},
 	)
 }
@@ -113,7 +116,7 @@ func FuzzCodecDBNodeCanonical(f *testing.F) {
 			codec := codec.(*codecImpl)
 			node := &dbNode{}
 			if err := codec.decodeDBNode(b, node); err != nil {
-				return
+				t.SkipNow()
 			}
 
 			// Encoding [node] should be the same as [b].
